Ui Compliance
feat(ui): harden ImGui compliance, docking lifecycle, and input capture

- add centralized UiInputGate and route viewport mouse/keyboard actions through it
- move Axiom Library to root-managed tool window ownership (with popout support)
- harden dock host/layout restore (NoDocking host + stable-frame debounce after minimize/restore)
- apply EditorConfig-driven DPI and multi-viewport settings at startup
- replace legacy direct io writes with io.Add*Event() input path
- add Dev Shell ImGui diagnostics (Metrics, Debug Log, ID Stack Tool, Begin-return loop toggle)
- contain AI worker-thread failures with try/catch and marshal results/status safely back to UI
- add hard-gate ImGui contract checker and wire it into CMake compliance targets
- remove duplicate visualizer source entry and clean project warning noise
- update docs in new structure: UI loop/state model + FAQ compliance matrix + docs index link

+++ app/src/Integration/GeneratorBridge.cpp
@@ -1,4 +1,5 @@
 #include "RogueCity/App/Integration/GeneratorBridge.hpp"
+#include <cmath>
 
 namespace RogueCity::App {
 
@@ -49,10 +50,10 @@ bool GeneratorBridge::validate_axioms(
             const auto& a = axioms[i];
             const auto& b = axioms[j];
             
-            const float dx = a.position.x - b.position.x;
-            const float dy = a.position.y - b.position.y;
-            const float dist = sqrtf(dx * dx + dy * dy);
-            const float combined_radius = a.radius + b.radius;
+            const double dx = a.position.x - b.position.x;
+            const double dy = a.position.y - b.position.y;
+            const double dist = std::sqrt(dx * dx + dy * dy);
+            const double combined_radius = static_cast<double>(a.radius + b.radius);
             
             // If distance < 20% of combined radius, it's extreme overlap
             if (dist < combined_radius * 0.2) {
